Dependency updates, and fixed that a few references where not correct. Also cleaned up some code in the memento implementation.

+++ src/Greenshot.Gfx/BitmapHelper.cs
@@ -920,7 +920,7 @@ namespace Greenshot.Gfx
         /// <param name="dpi">double with the dpi value</param>
         /// <param name="baseSize">the base size of the icon, default is 16</param>
         /// <returns>Bitmap</returns>
-        public static Bitmap ScaleIconForDisplaying(this Bitmap original, double dpi, int baseSize = 16)
+        public static Bitmap ScaleIconForDisplaying(this Bitmap original, uint dpi, int baseSize = 16)
 		{
 			if (original == null)
 			{
